feat: read/quaff floor items from description
Pretty straightforward change except that it was slightly tricky to get
the right item to use.  In the course of this I also refactored some of
the switch statements and generalized the quiver check -- you can quiver
from the description anything that is normally quiverable, rather than
using an item by item check. This required a somewhat ugly hack to keep
(q)uiver and (q)uaff from colliding, point even more towards the need to
overhaul this code in a more structured way.