Adjust the ammo generation probability
Is it because of Jiyva that the proportion of stones is high? If so, there is no reason to be high now, so I will lower it. (If I made a mistake, please correct it) Since throwing weapons are auxiliary weapons, it is reasonable that the supply and demand of bullets are difficult(But now it's too little. I'll raise it a little bit), but I think the supply and demand of bows and crossbows need to be smoother because they are main weapon.
I think it is ideal that the amount of ammo is somewhat insufficient when the player does not have the help of god.