feat: orbs of light
The first of several orb egos, an orb of light gives the player a radius
3 halo. Halo prevents invisibility and backlights the player (an extra
defensive penalty) in exchange for revealing invisible creatures and an
accuracy bonus. Alone the accuracy bonus isn't that big, but an ally
user can get quite a bit of a boon from halo.

This is probably the weakest of the orb egos, but that's ok. In general
orb egos will represent some kind of a tradeoff, as they're egos that
one can use in place of a shield without training.