fix: enable describe actions for item stacks
d40668a25 worked only on stacks of 1, because that was the only case
where the describe popup got access to the "real" item that was on the
floor. This commit expands things out so that for stacks of items if
they are at the player's feet, the real item is used all the way along
the chain. There was some required const munging that is not so great
(though there's a lot of code that did not seem like it should actually
be using const item_def * -- I started to work through this but the
domino effect made me back up.) This commit also makes describe actions
work for items in the full view describe that are at the player's feet.