feat: orbs of mayhem (nicolae)
The mayhem ego for orbs causes monsters who see the wielder kill
something to go frenzy. It only hits one monster per kill with a
coinflip for now. This could be adjusted up or down based on power, but
we definitely don't want to give out free Discord casts for a single kill.