Add a new L7 Hex: Enfeeble
Intended to give Hexes some more interesting utility to fill the gap
between the powerful early stabbing spells and Discord at level 8, Enfeeble
is a beam-targeted hex which irresistibly applies weakness and antimagic to
a single foe (even against infinite willpower). The target is also dazed
and blinded if they fail a willpower check (with a fairly substantial power
bonus).