docs: update placeholder README
I originally put this file here just as a test of how DGL setups would
handle subdirectories, but now that it has served its purpose, why not
do something with it. This overview is pretty brief and high level, but
as far as I know this kind of overview doesn't really exist anywhere
else.