feat: orbs of energy (bringing back the crystal ball)
This commit adds a new "energy" orb ego that offers passive channeling,
like Wucad Mu. The implementation is set to stack with Wucad Mu and any
other potential source of energy (currently none) in the following ways:

- The trigger chance is multiplied by the number of sources.
- The all sources must backfire for the backfire effect to trigger.

Wucad Mu's channeling was interesting, but compared to an enhancer staff
or other unrands the effect underwhelemed many players. In addition to
offering channeling with a shield, this commit buffs Wucad Mu by making
it count as two sources of channeling for internal calculations.

XXX: Unsure of the best way to communicate this to the player without
turning our inscriptions into a line of dubious code, inscribing it
*channel++ just doesn't sit right.