fix: temporarily revert changes to client.html template
This one gets me every time: templates update immediately even if the
server hasn't yet restarted. So a pre-refactor server trying to load
this template generates a 500 error because it doesn't have the new
config API. This commit adds a shim that will work until servers have
all restarted, and then it can be reverted.