feat: display potion/scroll rarity on the item knowledge screen
As a followup to 2f15ad60, this commit makes it easier to compare
potion/scroll rarities by displaying all of them on one screen.

When deciding which unidentified consumables to use in order to survive,
players often check the `\-` screen to see possible outcomes. So it
makes the item knowledge screen a perfect place for this information.