Make Formicid digging drain max health
The ability to dig outside of combat with no limit or meaningful cost has
been known to be a problem ever since Dig existed as a spell, and it
encourages a variety of undesirable behaviours. Formicids have been a
problem in this regard for a while (and even when the ability had a food
cost, this didn't appropriately limit its use).

To make digging have a noticeable cost, make it drain the user for each
tile dug, to make extensive out-of-combat digging hopefully much more
dangerous. To simplify the ability and make it slightly more useful as an
in-combat escape, also remove its movement speed penalty.