Store spellpower on cast for ongoing spells
Make Borgnjor's Vile Clutch, Searing Ray, Olgreb's Toxic Radiance, Flame
Wave, Searing Ray, Frozen Ramparts, and Dragon's Call all check spellpower
only once on cast, instead of continuously throughout the duration. Also
includes some minor cleanup for other spells that unnecessarily rechecked
power.

As noted in 9c71ebef145, this fixes their interactions with Sif's Divine
Exegesis, and for Searing Ray and Flame Wave it has no other meaningful
effect (since it's not possible to swap gear while channeling). For the
other spells, it does become possible to cast them with an enhancer and
switch to another item, but this is no different to the majority of other
spells with an ongoing component (such as transmutations, hexes, and
summons) where all that matters is the power at time of cast.