feat: add option autopickup_decide_on_sight
This is a weird option, but because some people relied on the old
behavior of how equipment would be marked as "needs autopickup" as soon
as you spot it, it's probably good to have. I've chosen the default for
the option to be what the current behavior was to not surprise anyone.