New unrand: Danyla's gloves
Continuing the 0.28 trend of new unrands, Danyla's gloves play in the
space of 'unrands for the early game'. The wearer is healed by drinking
unknown potions, and their MP is restored by reading unknown scrolls.
Is it wise to intentionally keep consumables unknown so that you can
use them as ersatz !hw and !magic? Maybe??

I considered also giving some bonus for reading ?id or some alternate
way to use up ?id with these gloves, but that seemed to go against the
design premise. Also, to be a bit too complicated.

More related changes to come in subsequent commits.