Move the Mace of Variability to the early game
It's often a trap to use the mace of variability later on, and there are
plenty of fun later-game unrands for m&f users to use. Let's switch this
back to a simple mace (from a great mace), buff its ench (mimicking klown
clubs), and put it in the early-game weapon set.