New early-game unrand: Throatcutter
Throatcutter is a +7 long sword of draining with a 33% chance to
instakill struck foes that are below 21 hp. Monster HP tends to
increase over the course of the game, so this is of course most
relevant early on. It's a bit like extra AC-ignoring damage, but
it only works when enemies are at low HP to begin with.

Terence gets a Maggie-esque 1% chance to spawn with this.