fix: bounds checks for monster lrd and ghostly sacrifice
This removes the hardcoded checks from 73776a4c32c7 and 4f857bc58ec3.
The reason for these crashes and for their variability was that the fake
monster used for the validity check is at position 0,0. So if the game
was saved on a level where a monster was in the upper left corner, the
iterators used in these two spells would fire a tracer at the monster
during game setup. The presence of a salt entrance in the save in #2308
was a red herring, it's just that this level had some monsters in the
right place.  It's probably still not ideal that the level even exists
at the point where this validity check happens?

Re-resolves #2308, should resolve #2281 (though I couldn't get an abyss
crash to happen on demand). I also added a bit more info to the ASSERT
that may make this easier to track down in the future.