Implement unrandart Club of the Night
Implements a +10 unrandart club with the following properties:
Stabs like a short blade
Gives stealth++
On melee kill, grants a decaying LOS reduction.
Each kill gives a ~25% reduction in line of sight, up to 75%.
This reduction decays by ~3% per turn.

This artifact was primarily created for the dumb pun.
However, it should be in line with the push for early game artifacts,
while having a unique and transformative effect
Being a club, it does not scale worth a damn.
Being a mace, it will not be ideal as a transitory item for enchanters.
On the other hand, the drastically reduced LOS with one or two successful
stabs allows for fun stab chaining or flight while the weapon is viable.