Rework shield penalties
To remove the hard skill thresholds for shield skill training, rework
shields to use a similar formula to armour encumbrance, with the penalties
being fully removed at 27 skill.

The new formula is:

2/5 * shield_penalty^2 / (20 - 3 * racial_factor) * (27 - shield_skill) / 27

where shield_penalty is 5 for bucklers, 9 for kite shields, and 13 for tower
shields, and racial_factor is -2 for large species and formicids, 0 for
medium species, 2 for kobolds, and 4 for spriggans.