Rework offhand punch checks
Simplify the check for offhand punches and make them overall slightly more
likely to trigger - don't make heavy armour reduce the chance of punching,
and get rid of Octopodes' improved chance for offhand punches (they still
the benefit of being able to punch with a shield/two-handed weapon). All
species instead have a 5% chance to trigger an offhand punch that scales
with unarmed combat skill to a 50% chance at max skill (compared to a a 33%
chance previously at 20 UC or higher).