refactor: simplify drop ui
Previously, the drop menu would let you drop to a prompt by hitting
enter with nothing selected. This is not very useful, and the code was a
mess; plus I plan to change the behavior of enter somewhat in a later
commit. (I guess there may be players out there who will need to remap
enter+letter to letter+enter at the worst?) If there's really demand for
the prompt mode it may be better to integrate it with prompt_invent_item
and get a working codepath for the toggle. But for now just make this
function much more straightforward -- as of this commit, enter with
nothing selected just exits.