refactor: wearing/removing armour
This refactors a good bit of the logic for taking off armour, as well as
a few other things in the inventory wear/unwear section.

I also removed a few TODOs, clarified the semantics of the return values
of the wear/unwear armour functions (they are not used anywhere anyway
[ Committer's note: they are used in clua to indicate success, but
 the inverted return values changed by this commit were not correct ]),
and added a few doc comments for some of the less intuitive functions.

[ Committer's note: Closes #2340. Rebased and fixed up, restored
  overzealously removed inscription checking prompts. ]