Remove the strength bonus from bite damage
It was the only attack remaining with a special-cased stat bonus on top of
the regular stat multipliers, and it was mostly an insignificant bonus.
Increase base bite damage by 1 for bite attacks (which does include VS bite
bite although this didn't previously benefit from the strength bonus), and
give Dragon Form an effective extra 2 levels of fangs since this would be
the main case where the strength bonus was actually noticeable.