feat: give an XP reward for the abyssal rune
With the abyss now stripped of most rewards, let's give the rune a
stronger incentive. It gives a potionlike effect for potions of XP
equivalent to two potions of XP (due to how its implemented it gives
slightly less skill XP than quaffing two !xp, but it's still a nice blob
of skill XP and two XLs).