Give XP for frenzied monsters' deaths
Frenzied monsters (from orbs of mayhem, Discord, or datura darts)
gave no XP due to a detail of their implementation. In general, it's
better for players to get XP for their kills (mostly for reasons of
'fun'), so let's fix that here. Balance issues are best handled through
other mechanisms.