fix: don't send crash milestones over webtiles sockets
This appeared in 6ef7baa362 but shouldn't be applied to the case of
crash milestones. (a) this code is super heavy and not safe after a
crash, and (b) it has timing issues with the socket closing, and so was
causing recursive crashes on a dead socket. (This is why no crash
milestones have appeared for the last couple weeks, they've all been
recursive crashes which don't generate milestones.)

To be clear, after this commit crashes will still be written directly to
the milestones file.