feat: add `bold_brightens_foreground=force`
Recently it came up that it can be helpful for console readability to be
able to use bold weight on bright colors even on modern terminals, not
just as a workaround; people (and by people I mean ebering) had been
doing this just by using 8 color mode. This commit provides an option to
not ignore bold_brightens_foreground in 16 color mode.