feat: don't grant explore piety to Jiyva worshippers in the Slime Pits
Granting exploration piety in the Slime Pits encourages Jiyva
worshippers to fully explore this threat-free branch.

A dozen tests show that if the player dives the Slime Pits and joins
Jiyva on the last floor, they'll have 63-105 piety from auto-exploring
the rest of the branch. The average is 86 piety, which is three stars.
Joining Jiyva before entering the branch guarantees that the player will
get to 100+ piety and reach four stars after fully exploring the Pits.