feat: reduce poison dart quantity
At current quantities a character can always have a mega ammount past
the early game, and is incentivsed to use them whenever they lack a
stronger ranged alternative at minimal cost. Let's lower the stack size
to put a bit more strategic pressure on their use.

Poison darts are in a bit of an awkward place relative to their other
thrown siblings, it might make more sense to move the poison brand to
boomers and have darts only be specialty items, but that's for a future
consideration.