rework orb of destruction damage formulas
Removes a secret hidden power stepdown and makes the formula linear. The new
formula is roughly equivalent damage to the old at low power, somewhat worse
at about 100 power, and very slightly worse at 200 power. Additionally, to
emphasize the positional aspect of the spell, change the damage dropoff at
low distance to more harshly penalize point-blank iood casts. Distance 1 now
deals a wimpy 30% damage (was 50%) and 60% at range 2 (was 70%). Range 3
and above are unchanged. Can be adjusted further if this is an overnerf.