feat: color/display diagnostic popup
The somewhat odd situation has emerged several times recently that
console players don't realize that their terminal is running in some
variant of an 8 color compatibility mode. First, when
allow_extended_colors=true was made the default and people thought 16
color mode was "new", and then more recently, when putty users have been
complaining about the new menu hover being too non-contrasty. (I'll
actually change that, but this isn't that commit.)

This popup provides in-game information that is mostly relevant to
console players about what their terminal and options are actually
doing. It's largely aimed to make it clear to putty users that something
is wrong, but should be more generally useful. For other build targets
most of the information isn't relevant, but it still shows the color
palettes corresponding to the 8/16 ANSI colors.

The popup is slightly hidden, under `?!`. I can imagine a later approach
where it is under an "advanced" submenu of the game menu, especially if
it eventually becomes an interactive popup. But for now this also lets
it be used at the main menu.

Still needs testing on windows.