fix: tone down default `may_stab_brand` (G-flex)
This was set to yellow, which by default on most terminals is extremely
bright as a background color, with poor contrast with many foreground
glyph colors.  Most players were either used to this and didn't notice,
or playing in 8 color putty where this comes out as `brown`, not the
brighter yellow that putty is also capable of displaying (but doesn't on
default settings). I think `brown` (which is typically a sort of mustard
color) has much better contrast on terminals I've tried anyways, and
it's not otherwise used for a monster highlight, so let's go with that.