Trim a floor from Lair
To continue shortening some of the branches which outstay their welcome, as
discussed in PR #1898 and partially implemented with 0.28's shortening of
Depths. More substantial changes have been discussed for the future (such
as only generating one S branch per game, and only one of Hell or Pan for
the extended endgame), but in the meantime shortening Lair slightly is a
good step in the right direction. Monster spawns and vault depths are
mostly adjusted such that the new Lair:1 is equivalent to the old Lair:2.