refactor: rename `.._brand` options to `.._highlight`
I assume there's some charming retro-roguelike and/or Australian
explanation for the use of "brand" here, but it seems completely
non-standard now and extremely confusing. I chose "highlight" as a much
more intuitive name. I also rewrote a fair amount of the options guide
for these beyond the name, partly with an eye towards including all the
search terms I initially tried when looking for this code the other day.
This is a full internal rename as well.

The actual implementation of these options is very odd and could use
some refactoring, but I'm not going to do that right now.

Needs testing on Windows (CI should be enough).