feat: option to disable arrow key menu control
I admit that 50% of the reason I'm adding this is so that the 1% of the
player base that is enraged by >90s UI features will stop bugging me.
But I suspect it could also be useful for accessibility purposes to be
able to disable hover in console -- for example I doubt it works well
with screen readers without some specific effort that I don't really
know how to do.  As a side effect this option does disable arrow control
for all menus that had it added in the 0.27 cycle.