fix: treat Passwall like stair and shaft-self delays
Treating Passwall like a usual delay, as opposed to a multi-turn
move commitment like shaft self and stairs, created all sorts of weird
things as documented in Mantis #12239, #11825, and #10961.

Treating it like a stair or shaft self delay makes it behave more like
"movement" and gives it a consistent set of rules.