Allow upward stone stairs to be used for LRD
Swamp can be a difficult place in which to find any valid LRD targets.
For those who play in tiles mode, it can be confusing that up-stairs
aren't targets for LRD; it's clearly a stone structure and is described
as such as well. Down stairs on the other hand, are a hole in the floor.