feat: Update a Depths float vault
For `minmay_skeletal_surprise`, use a better substitution setup and add
more Depths-appropriate nasties monsters. Namely mix in a small chance
for ancient/dread liches, keeping averages the same. For the nastier
monsters that place in the gold rooms, place a superb or star item
instead of just more gold.