feat: Update a Depths mummy-themed vault
For `lemuel_minitomb`, update the vault to monsters that are appropriate
challenges for Depths, and revamp the appearance, layout, and loot.
Don't use plain mummies, using spectral things at the low end, along
with guardian mummies and mummy priests to go along with the single
royal mummy. Replace some of the gold in the final chamber with super
and star item loot, which is a more meaningful award. Use tomb-theme
wall and floor tiles, and make the layout more favorable to monsters in
the final two chambers. Allow the vault to place in all of Depths and
Crypt.