Option to ignore all items of a type during autoexplore
Updates the 'cannot pick up item, ignore Y/N' autoexplore prompt

Currently relies on the 'yesnoquit' function from prompts.cc which has not been otherwise used in ~5 years, and seems to lack much of the functionality of yesno.

This causes the following issues:
Yes and Always both cancel autoexplore, against intended behaviour.

The prompt offers a redundant quit option with identical behaviour to the no option.

May fail gracelessly with HUPs as YNQ does not seem to have a mechanism for handling them the way YN does.

May not function with non-keyboard UI schemes

May not function with LUA inputs