feat: stash search menu improvements
The ultimate aim of this is to make this menu workable with arrow key
selection. Previously, carried item results weren't hotkeyed at all
(presumably because travel is pointless), but this didn't quite work for
arrow selection.

* All stash search results are selectable.
* Visually, items at the player's feet and in their inventory are
  grouped together, and the former are now clearly indicated in the
  menu. A hotkey (,) quick-moves the menu between these sections.
* In travel mode, selecting an item at the player's position doesn't do
  anything; previously this was true for carried items (because they
  weren't selectable) but did open for things on the ground. Possibly it
  would be clearer to do something in this case, maybe a message? Or
  open a description? These items are also greyed out in travel mode.
* Subheads are now cyan for better readability
* Various supporting Menu API improvements (including some bits that are
  probably extremely useful in the future but didn't end up fully needed
  here.)

It would be possible, but I found it a bit confusing on first
implementation, to start the menu with the first item not at the
player's position selected. Will revisit this on feedback. I think with
the , hotkey this isn't as important.

Other TBD improvements: the keyhelp could be improved and standardized,
I think it might be good to support right click to describe in this
menu.