fix: let webtiles mouse dungeon clicks interrupt travel/repeats
Previously, the click was just discarded during interruptible
travel/repeats. This change will let the fake key enter the key handling
queue, and `kbhit()` return true, which is the main check for
interrupting such things. I think it's ok to return a keycode
unconditionally for click handling, and CK_MOUSE_CLICK is innocuous, but
we'll see.