fix: make canvases non-selectable
On touchscreens (ios), it is a bit too easy to long tap and select the
canvas, which is then somewhat annoying to undo. This commit isn't a
real solution because it is still possible to long top the canvas, but
it prevents the canvas itself from being selected. (In my testing, some
very thin element to the left of the canvas gets selected?) This means
that if you tap the canvas itself the selection disappears.