feat: give out some fancy shields in Snake
Give Naga Warriors a chance of getting a good_item shield. Overall this
will result in an average of 2-3 good shields per Snake.

With the lootfeel shifts of Spider and Swamp, and Shoals' throwing
items, Snake was already somewhat distinguished by the chance of getting
a kite or tower shield off of a Naga warrior. Adding a small chance for
the shield to be fancy should make this pop a bit more as lootfeel.