fix: exert more control over iOS virtual keyboards
The behavior is still somewhat flaky (a few bugs noted in comments), but
this commit brings improvements. Basically, it tries to never let
anything else but the mobile input box have focus, while the keyboard is
displayed -- this is apparently what it takes on iOS to keep the
keyboard up. Once it's down, focusing a text input doesn't bring it back
up, and despite trying many recipes, I could not find a way to
programmatically bring it back up.  This prevented several alternative
approaches that I considered, e.g.  using a timeout. The focus control
involved here is quite aggressive and hopefully won't break anything on
Android.

This also does some better styling of the mobile input box, rather than
relying on the browser focus highlight.