fix: tweak webtiles mouse hover behavior with multiple popups
Previously, if you open a popup over a menu, and closed the popup, the
mouse position at the time would determine hover, which can be a bit
confusing (especially big menus where the mouse is likely to be
somewhere in the menu, and especially for arrow key selection). This
keeps hover on the previously hovered item (if any) when focus is
returned. We'll see if something even more fine-grained would be better,
e.g. suppress hover if previous item was not selected by mouse. It's
also still TBD under what conditions if any a pure mousemove (with no
leave or enter events) should set the hover, right now it doesn't.