fix: tweak console hover rendering on yes/no prompts
There is a difference in how hovers render for formatted vs
non-formatted menus; this basically results in the hover being the width
of the menu for non-formatted ones. Since yes/no menu prompts are pretty
variable, this can sometimes be odd (and it's worse if someone overflows
the terminal width). So, as a quick workaround, just mark this with
MF_ALLOW_FORMATTING.