feat: add "always" prompting to yesno
There's one place in particular where we ask the player a specific
question about a one-time situation that we can generalize both
technically and sensibly for ui: "Could not pick up an item here"
prompts.

There may be others, so I've implemented this as a way to prompt for
either a one-time or global change.

Previously, a general [A]lways prompt was available
via yesnoquit, but that has been unused for 6 years, has a FIXME for
being a mess, and is missing some updates and modernizations of yesno.