feat: removed unused yesnoquit
It was partially provided to the lua API in 2008 when implementing toll
zigurats for prompting if the player wanted to pay. Or at the same time,
that ended up using yesno. Subsequently it slowly rotted against the
weight of the FIXME: looking to be a real mess and disuse in the main
codebase. This might be a pain point for some clua users, but yesno is
quite full featured and ternary decisions (yes/no/quit) can be factored
into nested binary decisions by those downstream users.