fix: linebreak Y/N popup prompts (ebering)
It is very unclear to me whether adding automatic linebreaking would be
a good idea to MenuEntry items (which are used for many things), so I
added a convenience function that only does this happen on demand, and
used it for `yesno` calls. N.b. it works fine on Menu titles, but
console regular menu entries do not work with multiline text.

I also did a bit of relevant MenuEntry refactoring.