feat: Update a vampire-themed float vault
For `lemuel_vampire_tower`, revamp the vault to have better monsters
that scale with depth of placement.

We remove the down-hatch, since it served little purpose other than
limiting placement. Move the guaranteed altar to where the down hatch
was, and put a superb_item on the altar to go along with the other two
loot items. Allow the vault to place on the final levels of Dungeon,
Depths, and Crypt.

Rework the monster tiers to follow a more logical progression, using
monsters appropriate for placement depth. For the vampire boss, place a
vampire for earlier Dungeon, upgrading to a vampire knight, with a
placement of either Jory or a vampire knight decked out with thematic
equipment of high quality for Crypt and Depths.