fix: reentrant calls to show_map via descriptions
When viewing a staircase description, you can now use ]/[ in the
description popup to go right to the map view on the other side. Since
description popups are possible in `X` view, this could lead to weird
stacking of map view UIs that were really not designed to happen. There
was also a (possibly independent, but I'm not entirely sure) set of bugs
where exiting the describe popup wasn't handled right and made the
player travel to the stair location. This commit fixes all of this by
refactoring `show_map` to allow graceful reentry without creating a new
map view layer. Well, the behavior is graceful, the code isn't, but
it'll do for now.