fix: prevent pickup of non-env items via describe actions
If an item_def is otherwise valid but is a copy of a real item in
`env.items`, and the player is in the right position, this code would
allow trying to pick it up and then crash. Add a check for a valid
index. (The logic here is patterned off of `invalid_monster_index`.) The
immediate effect is on descriptions accessed via stash search, and it
would be better to actually allow the (g)et to work in this case, but it
may be somewhat unavoidable that stash search is working with copies, so
getting (g)et to work will be a bit of a project.