fix: select first travelable item in stash search
Some players use this menu as a quick shortcut while collecting items
that they are sure already exist, and I guess having to hit ','+enter is
too much work. This lets enter immediately travel to the first
travelable item.

I'm undecided as of yet what to do with the lettering: I respect that
many players have some muscle memory for ctrl-f a, but on the other
hand, I don't think the fact that a kind of dodgy UI decision is
enshrined in people's muscle memory is a good reason to keep it. I'm
definitely opposed to making items on the current position fully
unselectable in this UI, as was done (for carried items) in response to
https://crawl.develz.org/tavern/viewtopic.php?f=8&t=20374. This makes
the whole UI less usable and there's a ton of players who don't need or
want the reversion that was (rather unconstructively) advocated for in
that old thread.