fix: default `easy_floor_use` to false
This option applies the action (read, wield, etc) to the floor item
immediately on pressing ',' in the relevant menu if there is exactly one
item of the right type on the floor. This behavior is really
inconsistent with how menus work in general, and with what ',' does when
there's more than one item. I'm keeping the option since it's been a
default for a very long time. If people really don't like this new
default I'm open to doing something further; one idea might be that if
there's one item, assign it an actual (non-comma / non-letter) hotkey.