fix: clean up some messy use item menu code
This simplifies the use item menu so it doesn't have to entirely rebuild
the menu every time the section changes. Rather, this commit adds an
enable/disable toggle at the MenuEntry level that suppresses all hotkeys
when disabled, and then changes this for relevant entries when switching
sections. It also generalizes the section-switching code using the newer
cycle_headers pattern.

In addition, a minor visual tweak: indent non-hotkeyed items in these
menus so that there horizontal position doesn't change on section
change.