feat: quiver menu lettering improvements
A long-time complaint about the newquiver menu is that the hotkeys are
fairly arbitrary. This is challenging to resolve, since things that
appear in this menu come from items, spells, and abilities, and there
are likely to be collisions past the most basic of characters. A lot of
players with this complaint have their preference for which lettering is
the most important (I think that team item is the most vocal, because
the old quiver menu just used item lettering), but these can't be
mutually resolved.

This commit attempts to use some of the fancier technology from
UseItemMenu to resolve this problem. It adds a new mode to the quiver
menu that focuses in on one of the three categories and uses the
lettering from that category; the UI improvements from UseItemMenu
hopefully make it easier to switch between the categories (e.g. arrow
keys move smoothly between them, mouse hover changes categories*, and
there's a toggle key as well. Focus mode on/off is implemented via a
persistent option, so while you can manually set the rc option, it just
remembers whatever mode you last had it in. Within a single play session
it also remembers the section while in focus mode. So a typical use case
for someone who doesn't quiver anything except items would be to hit !
to put the menu in focus mode, and not think about it ever again (until
they upgrade, I suppose).

This commit also involved a bunch of related fixes for stuff in Menu and
UseItemMenu, as well as moving the quiver menu keyhelp to the footer.

[*] n.b. there's been a regression somewhere along the line where this
stopped working on local tiles, I'll deal with that separately.