fix: make the unwield menu item usable
Previously, this menu item was basically cosmetic, and didn't have a
hotkey, preventing it from being hovered (in local tiles at least). This
commit turns it into a real menu item that can be selected. While I was
at it I improved some of the visuals of this item -- it uses
form/species custom unarmed attack strings (though a lot of forms simply
prevent this menu altogether), and indicates visually if unarmed is
currently chosen.

To match previous behavior, this menu starts with the first real item
selected. We'll see if this is confusing or not.