fix: explicitly set UI_NORMAL before load_level in webtiles
Messaging from portals happens during load_level, but on new char
creation, UI_CRT is set by the character selection popups, and wasn't
cleared in webtiles until load_dungeon. Since this state change is what
is required for the initial show() call to the `normal` div in webtiles
for this case, a more triggered during load_level would still wait for
input but with the entire game div invisible. This is what led to the
delver bug when landing on a level with a portal. Arguably it is also a
bug that this messaging happens so early on initial load, but this seems
extremely hard to change (marker init can't be moved later without what
I seem to remember is complete disaster, so it would require deferring
marker init messaging somehow).

Resolves #2318, mantis 12343.

(cherry picked from commit b0e57e52b49aa929b11540c464f9f70d613b747a)