fix: reset scroll on spell search
If the scroller was sufficiently far down, a search could result in no
menu lines visible in console because the scroll didn't get adjusted. I
haven't quite figured out why the Scroller m_scroll sanitization, which
appears to be correctly implemented, doesn't kick in until a manual
scroll. Maybe there's a missing _allocate? In any case, this commit
explicitly manages the scroll for spell search, trying to keep it on the
same spell if possible, but otherwise resetting to the top.

Resolves #2303