fix: suppress mores during describe actions
This prevents graphical glitches in console and a hard-to-spot key input
prompt in webtiles.  Just suppressing the more is not an ideal solution
as it can make you miss force_mores, but I'm not sure there's an
alternative that wouldn't take some major refactoring. (I could imagine
an RAII object that instead of suppressing output, collected it in a
message tee and showed a popup if there was a force_more. But this is
definitely in non-freeze territory.)

Resolves #2232

(cherry picked from commit 022ca4c89056c948c7f6d2dc3dc46d38cef04442)