fix: quivering ammo from item description
This call needed to correctly handle the ammo quiver (and the inscribe
check), and to avoid code duplication I pulled similar relevant code out
of ActionSelectMenu into its own call. Overall, this whole setup could
possibly use some simplification at this point, it was written to
support arbitrary numbers of quivers but I am not sure that's ever going
to be realistic.

Resolves #2257