fix: adjust some casing in mut short names
These short names are being used as popup titles for `A` screen popups,
and given this it is better to systematically uppercase relevant
abbreviations. Because this field was designed for the % screen list,
there was mostly a lowercase convention for these fields (with a few
exceptions), but it doesn't seem to matter in any technical way if this
casing is adjusted; db lookup is case insensitive. They look ok in the
short list too, and a few things were already cased there.

(Also because these have been designed for brevity, they often don't
work well as titles, and I think in the long run it may be good to have
an explicit "title" field of some kind in this data that allows writing
a better one. But that's a bit of a project.)