feat: Updates to octagon-themed float vaults
For minmay's set of octagon vaults, namely `minmay_shoctagon`,
`minmay_roctagon`, `minmay_firetagon`, and, yes of course,
`minmay_coctagon`, updates to placement, monster sets, loot, and syntax.

These vaults all have a lower-tier monster set that places 8 monsters
on-average along with a single out-of-depth "boss" monster, but the
monster sets don't change much with depth. I've updated the maps to have
appropriate monsters for earlier Dungeon (either D:10-D:12 or D:12-D:13,
depending on the map), late Dungeon, Elf, and Depths. For Depths, all
vaults can now place on every level of that branch. I've replaced
`ident:type` with `ident:all` in all cases.

Simplify the loot/monster placement logic to not have a rare chance to
add a lot more of either monsters or loot. Nearly doubling the loot
doesn't work well in a vault where the loot is already generous and
every loot item is thematic. Remove these rare substitutions in the
common setup function. Tweak the NSUBST to place a few more monsters so
we keep the same average.

For the unrand weapon item, follow the lead of `minmay_coctagon` and
have thematic randart weapons of the same weapon class but of subtypes
different from the unrand. This way we're not placing e.g. the Arc Blade
100% of the time, like a guarded unrand vault. Also add to these vaults
some other thematically appropriate unrands with randart alternative to
these vaults in cases when unrands exist.